pyproject.toml
[project]
name = "pradotube-scripts"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "supabase",
    "pyyaml",
    "requests",
    "boto3>=1.42.74",
    "python-dateutil>=2.9.0.post0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
]

[project.scripts]
sync-subscriptions = "scripts.sync_subscriptions:main"
sync-downloads = "scripts.sync_downloads:main"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
